Fix keystone provider to work with v3 auth_url
Fuel-library commit: https://review.openstack.org/#/c/207548/
This patch was not upstreamed because issue was fixed
in another way using of new openstackclient. See comments:
https://review.openstack.org/#/c/207873/

Python-openstackclient can't be now updated, so we need
to merge this patch as is, and revert it later after
switching on Liberty packages and upstream solution.

Change-Id: I6afd35bfbf81980504efb5ac23762b22bf4cd064